chore(20076): Generic clean up for SequentialThreadsTaskScheduler threads in test
