+++ platform-sdk/swirlds-component-framework/src/test/java/com/swirlds/component/framework/model/DeterministicModelTests.java
@@ -19,6 +19,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import com.swirlds.base.test.fixtures.time.FakeTime;
 import com.swirlds.base.time.Time;
 import com.swirlds.common.metrics.noop.NoOpMetrics;
+import com.swirlds.component.framework.schedulers.SequentialTaskSchedulerAliveThreadCleanup;
 import com.swirlds.component.framework.schedulers.TaskScheduler;
 import com.swirlds.component.framework.wires.input.BindableInputWire;
 import com.swirlds.component.framework.wires.input.InputWire;
@@ -35,7 +36,7 @@ import java.util.function.Function;
 import org.hiero.base.utility.NonCryptographicHashing;
 import org.junit.jupiter.api.Test;
 
-class DeterministicModelTests {
+class DeterministicModelTests implements SequentialTaskSchedulerAliveThreadCleanup {
 
     /**
      * Build a handler method for "components" in the wiring generated by generateWiringMesh(). The goal of these
@@ -449,40 +450,43 @@ class DeterministicModelTests {
 
         model.start();
 
-        int expectedCountA = 0;
-        int expectedNegativeCountA = 0;
-        int expectedCountB = 0;
-        int expectedCountC = 0;
-        int expectedCountD = 0;
+        try {
 
-        for (int i = 1; i < 1000; i++) {
-            channelToA.put(i);
+            int expectedCountA = 0;
+            int expectedNegativeCountA = 0;
+            int expectedCountB = 0;
+            int expectedCountC = 0;
+            int expectedCountD = 0;
 
-            expectedCountA = hash32(i, expectedCountA);
-            expectedCountB = hash32(i, expectedCountB);
-            expectedCountC = hash32(i, expectedCountC);
-            expectedCountD = hash32(i, expectedCountD);
+            for (int i = 1; i < 1000; i++) {
+                channelToA.put(i);
 
-            if (i % 7 == 0) {
-                expectedNegativeCountA = hash32(-i, expectedNegativeCountA);
+                expectedCountA = hash32(i, expectedCountA);
+                expectedCountB = hash32(i, expectedCountB);
+                expectedCountC = hash32(i, expectedCountC);
+                expectedCountD = hash32(i, expectedCountD);
+
+                if (i % 7 == 0) {
+                    expectedNegativeCountA = hash32(-i, expectedNegativeCountA);
+                }
             }
+
+            int maxTicks = 10_000;
+            while (true) {
+                if (maxTicks-- == 0) {
+                    fail("Model did not quiesce in time");
+                }
+                model.tick();
+                if (expectedCountA == countA.get()
+                        && expectedNegativeCountA == negativeCountA.get()
+                        && expectedCountB == countB.get()
+                        && expectedCountC == countC.get()
+                        && expectedCountD == countD.get()) {
+                    break;
+                }
+            }
+        } catch (Exception e) {
+            model.stop();
         }
-
-        int maxTicks = 10_000;
-        while (true) {
-            if (maxTicks-- == 0) {
-                fail("Model did not quiesce in time");
-            }
-            model.tick();
-            if (expectedCountA == countA.get()
-                    && expectedNegativeCountA == negativeCountA.get()
-                    && expectedCountB == countB.get()
-                    && expectedCountC == countC.get()
-                    && expectedCountD == countD.get()) {
-                break;
-            }
-        }
-
-        model.stop();
     }
 }
